refactor(web): STRUCT-9 finish - extract Library and BookDedup sub-components
Extract 7 sub-components to reduce file sizes and improve maintainability:

BookDedup.tsx (3425 lines) -> 4 components:
- DedupAdvancedScanTab.tsx: advanced scan controls tab
- DedupAuthorTab.tsx: author deduplication tab
- DedupSeriesTab.tsx: series deduplication tab
- DedupReconcileTab.tsx: reconciliation tab

Library.tsx (3243 lines) -> 3 components + shared types:
- LibraryToolbar.tsx: sticky toolbar (batch/library-actions modes)
- LibraryBookGrid.tsx: book grid with pagination, filters, soft-deleted section
- LibraryDialogs.tsx: all 22 dialogs and import-paths list
- libraryTypes.ts: shared types (ImportPath, BulkActionResult, etc.)

Dead code removed: hidden <Paper display:none> batch actions section
(~250 lines, superseded by BatchToolbar)
TypeScript: 0 errors with noUnusedLocals + noUnusedParameters
